Addon store: Invalidate cache based on NVDA API Version (#14685)
Part of #13985

Summary of the issue:
Within NVDA the add-on store keeps a cache of data from the addon store API. This is invalidated after 6 hours.
This cache should also be invalided if the version of NVDA changes. The list of available add-on will change for different versions of NVDA.

Description of user facing changes
The cache should be updated when the NVDA version changes.

Description of development approach
Add the current NVDA API version when the fetch occurs to the caching schema.