soffice: Report page-relative caret position when available
Fixes #11696
Related LibreOffice bug report:
https://bugs.documentfoundation.org/show_bug.cgi?id=136760

Summary of the issue:

When reporting the review cursor location in LibreOffice Writer,
the announced pixel-related coordinates are not particularly useful
to determine where in the document/page the cursor is currently located.

Description of how this pull request fixes the issue:

Similar to how it is done for Microsoft Word, report
the current cursor/caret location relative to the current
page when that information is available.
To retrieve that information, this introduces and makes use of
a new `SymphonyDocumentTextInfo` class that retrieves the
page-relative cursor position via accessible attributes from the
document object.
These attributes are added to LibreOffice in this corresponding
LibreOffice change:
https://gerrit.libreoffice.org/c/core/+/149051
The values returned by LibreOffice Writer are in twips, so
convert that to inches or centimetres, depdending on whether
or not imperial measurement units are expected.
If the attributes are not available, fall back to the handling
that was already used so far.

Testing strategy:
* use a LibreOffice build that includes the changes from
  https://gerrit.libreoffice.org/c/core/+/149051
* start LibreOffice Writer
* trigger reporting the review cursor location by pressing
  NVDA+shift+numpadDelete
* verify that the current caret position is announced
  in centimetres or inches from the left and top edge of the
  page