Settings Ring: Lazily load available setting values the first time they're needed (#14704)
A settings ring entry for a String setting caches its possible values when creating the StringSynthSetting object. This can be costly if it takes some time for the possible values to return, for example when they're coming from a remote system in my remote desktop add-on.

Description of development approach
Load and cache the available settings values the first time they're fetched. This shouldn't have any noticeable impact for most synths.that