HID Braille Driver: add more mappings for arrow keys (#14713)
The Standard HID Braille driver does not provide mappings to the arrow keys as expected.
Currently the joystick can be used for arrow keys and enter, but not the dpad. Some Braille devices, such as the Orbit Reader, only have a dpad and no joystick.
Also, currently space+dot1 and space+dot4 are mapped to nextLine and previousLine braille commands. Manufacturers would prefer that these gestures map to up and down arrow, so that they also work in lists and on the Desktop etc.
Description of changes
When using a Braille Display via the Standard HID braille driver, the dpad can be used to emulate the arrow keys and enter. Also space+dot1 and space+dot4 now map to up and down arrow respectively.