Fix #14741 (#14744)
In #14704, we introduced lazy loading of some settings ring settings. However, this resulted in error sounds being played when moving in the list of OneCore voices.

Description of user facing changes
Fix of the regression.

Description of development approach
The max value of a synth setting is based on the number of available settings. This is now dynamically fetched.