Allow exporting a gesture map without saving it (#14582)
Summary of the issue:
You can currently load a gesture map from a file or a dictionary, but you can only save it as a file, i.e. you can't go from gesture map to the flattened dictionary format.

Description of user facing changes
None

Description of development approach
Split out the save function into another function called export. The save function uses the export function internally. This can also be used for unit testing the code as demonstrated. Furthermore, it is useful for braille display driver authors to develop a gesture map in the user gesture map, then export it from the python console to include it in their driver in the proper format.