User guide: clarify the usage of character level in the Punctuation/symbol pronunciation dialog (#14737)
Summary of the issue:
People are
In Punctuation/symbol pronunciation dialogue, in the level field, there is an options named "character". People are wondering what this option does since it is not documented in the user guide.

Description of user facing changes
Added a sentence in the doc to describe it.