When reporting URL of a link use caret position instead of navigator object (#14723)
Closes #14659

Summary of the issue:
In #14583 @XLTechie added a global commands which allows to report a URL for the link at the position of the navigator object. While this is a huge improvement compared to either having copy the URL to the clipboard or accessing the URL from the status bar from the web browser, the fact that these scripts report destination for the link at the navigator object rather than focus / caret position makes their usage impractical for people who prefer to work with review cursor / navigator object not following focus / caret.

Description of user facing changes
When user requests URL of a link NVDA reports URL of a link at the position of the caret / focus.

Description of development approach
Script for reporting URL was modified so that it first tries to get an object at the current caret position, and if that fails from the focused object. User guide and documentation for global commands were updated accordingly.
After @XLTechie's review it was noticed that the script fails to work for links without href attribute, as part of this PR I've added handling for such links.