PR template: indicate when PR state need to be set to 'Ready' (#14758)
See #14610 (comment),

Summary of the issue:
Sometimes, people open PRs as a draft as required but do not seem to know that they need to transition it to "Ready for review" so that someone at NVAccess look at it.

In the PR template we can read:

Please initially open PRs as a draft. See https://github.com/nvaccess/nvda/wiki/Contributing

But nothing on when it should be set "Ready for review".