Bugfix: display ampersand character in the message of `nvdaControls.MessageDialog` (#14607)
Summary of the issue:
When a message containing ampersand has to be displayed in nvdaControls.MessageDialog, the ampersand character is interpreted as an indication for accelerator key. Thus the ampersand character is not displayed and the following character is underlined.
Putting here an accelerator key does not make sense here since the wx.StaticText used in this dialog is not the label of a focusable element.

The real-life situation where this issue occurs is when installing NVDA Dev & Test Toolbox. The dialog indicates "Extension : NVDA Dev Test Toolbox 3.0" (without ampersand) instead of "Extension : NVDA Dev & Test Toolbox 3.0" (with ampersand).

Description of user facing changes
Messages displayed in nvdaControls.MessageDialog will be displayed correctly when the message contains ampersand character. E.g. add-on install dialog and add-on install errror dialogs will display correctly the name of add-ons containing ampersand character in their name.

Description of development approach
Doubled the ampersand character when present in the message.