Notepad with tabs: speak when switching between tabs (#14588)
Closes #14587

Summary of the issue:
In recent notepad releases on windows 11 insider, notepad has gained the ability to open multiple files in the same window with tabs. When switching between tabs with control tab, NVDA does not the speak the currently selected tab or it has changed at all.

Description of user facing changes
When CTRL tab is pressed, the new tab name will be spoken, similar to web browsers.

Description of development approach
Implement behavior for UIA_elementSelected, if the object is a tab and is selected, speak it.