added toLWTH conversion if startLocation is rectLTRB (#14671)
Fixes #12424

Summary of the issue:
JABTextInfo can return rectLTRB which is then used in the boudingRects property. this can be an issue for functions that expect rectLTWH types to be returned.

Description of user facing changes
None

Description of development approach
walked through where JABTextInfo could return a rectLTRB instance. found it