Updated the doc for the scratchpad option (#14788)
Closes #14773

Summary of the issue:
In #14773, the user wanted to document when the code is loaded from the scratchpad directory.

Description of user facing changes
As discussed in the issue, it is hard to add more details in the GUI. Thus, the User Guide's paragraph describing this option has been updated. Not that it is easily reachable from the GUI thanks to context help.

The following updates have been made:

Added a sentence to indicate when all the modules from the scratchpad are loaded
Added visual enhancement providers to the list of the modules that can be loaded from the scratchpad.
Removed the sentence comparing the current situation with older versions of NVDA when the scratchpad did not exist and test modules were directly loaded from appModule, globalPlugins, etc. subfolders. Indeed, the scratchpad has been introduced in NVDA 2019.1, which is now quite old.