Added a Development Guide chapter describing and listing extension points (#14810)
Closes #14648

Summary of the issue:
There is currently no definitive list of all available extension points. Since these seem to be the preferred future direction for add-ons and other code to interoperate with NVDA internals, it would be valuable to include a list of them in the Developer Guide.

It would also be valuable to include various examples and further documentation there, but that can come in a later PR.

Description of user facing changes
None.

Description of development approach
Created a new chapter in the Developer Guide, describing very briefly what extension points are, mainly as a concept introduction. This could stand to be expanded significantly.
Used grep to find all occurrences of "extensionPoint" in the source.
For each one, determined why the module was being imported, and if it was actually used to create an extension point, added appropriate entries to the dev guide describing that extensionPoint.