Display in braille the character number upon request (#14837)
Fixes #14826

Summary of the issue:
When pressing 3 times numpad2, the character number is reported by speech (both in decimal form and hexadecimal form). However, for braille users it would be useful to have this information.

Description of user facing changes
When pressing three times numpad2 to call the script "Report current character in review", the reported information (character number in decimal and hexadecimal) is also reported in braille, not only by speech.
First press and second press of this script remain unchanged since:

First press only report the current character which is already available at the review cursor's position on the braille display, provided that the braille follows automatically (default settings) or review.
Second press is a feature dedicated to speech users, when they cannot hear correctly the character name.
Description of development approach
Added a braille update as done in ui.message. We do not use directly ui.message since the hex value has to be spelt.